Split large components into smaller, human-friendly subcomponents for WorldMap, ScoreboardModal, and PersistentGameHeader. Modularized logic and UI for maintainability.
