+++ app/components/WorldMap/index.ts
@@ -1,0 +1,7 @@
+export { default as WorldMapContainer } from './WorldMapContainer';
+export { default as MapMarkers } from './MapMarkers';
+export { default as TargetCityMarker } from './TargetCityMarker';
+export { default as ProvisionalMarker } from './ProvisionalMarker';
+export { default as useLeafletMap } from './useLeafletMap';
+// For compatibility, also export WorldMap as WorldMapContainer
+export { default as WorldMap } from './WorldMapContainer';
